Refactor - remove useSessionToken hook.
