--- src/applications/check-in/hooks/tests/useSessionToken/useSessionToken.unit.spec.jsx
@@ -1,87 +1,0 @@
-import React from 'react';
-import { expect } from 'chai';
-import sinon from 'sinon';
-import { render, fireEvent } from '@testing-library/react';
-
-import TestComponent from './TestComponent';
-
-describe('pre-check-in', () => {
-  describe('useSessionToken', () => {
-    describe('getCurrentToken', () => {
-      it('window is null', () => {
-        const window = null;
-
-        const component = render(<TestComponent window={window} />);
-        const button = component.getByTestId('get-button');
-        fireEvent.click(button);
-        expect(component.getByTestId('from-session').innerHTML).to.equal('');
-      });
-      it('calls getItem', () => {
-        const getItem = sinon.spy();
-        const window = {
-          sessionStorage: {
-            getItem,
-          },
-        };
-
-        const component = render(<TestComponent window={window} />);
-        const button = component.getByTestId('get-button');
-        fireEvent.click(button);
-
-        expect(getItem.called).to.be.true;
-        expect(getItem.calledWith('health.care.pre.check.in.current.uuid')).to
-          .be.true;
-      });
-      it('key is not found', () => {
-        const window = {
-          sessionStorage: {
-            getItem: () => null,
-          },
-        };
-
-        const component = render(<TestComponent window={window} />);
-        const button = component.getByTestId('get-button');
-        fireEvent.click(button);
-        expect(component.getByTestId('from-session').innerHTML).to.be.empty;
-      });
-      it('key is found', () => {
-        const window = {
-          sessionStorage: {
-            getItem: () => JSON.stringify({ token: 'some-token' }),
-          },
-        };
-
-        const component = render(<TestComponent window={window} />);
-        const button = component.getByTestId('get-button');
-        fireEvent.click(button);
-
-        expect(component.getByTestId('from-session').innerHTML).to.equal(
-          JSON.stringify({ token: 'some-token' }),
-        );
-      });
-    });
-    describe('setCurrentToken', () => {
-      it('saves data to name spaced key', () => {
-        const setItem = sinon.spy();
-        const window = {
-          sessionStorage: {
-            setItem,
-          },
-        };
-        const testToken = 'testToken';
-        const component = render(
-          <TestComponent window={window} token={testToken} />,
-        );
-        const button = component.getByTestId('set-button');
-        fireEvent.click(button);
-        expect(setItem.called).to.be.true;
-        expect(
-          setItem.calledWith(
-            'health.care.pre.check.in.current.uuid',
-            JSON.stringify({ token: testToken }),
-          ),
-        ).to.be.true;
-      });
-    });
-  });
-});
