+++ src/applications/check-in/pre-check-in/containers/withAuthorization.jsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 import { makeSelectCurrentContext } from '../selectors';
 
 import { useFormRouting } from '../hooks/useFormRouting';
-import { useSessionToken } from '../../hooks/useSessionToken';
+import { useSessionStorage } from '../../hooks/useSessionStorage';
 
 import { SCOPES } from '../../utils/token-format-validator';
 import { URLS } from '../utils/navigation';
@@ -16,7 +16,7 @@ const withAuthorization = Component => {
     const { token, permissions } = useSelector(selectCurrentContext);
 
     const { jumpToPage, goToErrorPage } = useFormRouting(router);
-    const { getCurrentToken } = useSessionToken();
+    const { getCurrentToken } = useSessionStorage();
 
     useEffect(
       () => {
