+++ src/applications/check-in/pre-check-in/pages/Landing/index.jsx
@@ -9,7 +9,6 @@ import { createInitFormAction, createSetSession } from '../../actions';
 
 import { useFormRouting } from '../../hooks/useFormRouting';
 import { useSessionStorage } from '../../../hooks/useSessionStorage';
-import { useSessionToken } from '../../../hooks/useSessionToken';
 
 import { createAnalyticsSlug } from '../../../utils/analytics';
 import { createForm, getTokenFromLocation, URLS } from '../../utils/navigation';
@@ -36,8 +35,7 @@ export default function Index(props) {
 
   const { router } = props;
   const { goToErrorPage, jumpToPage } = useFormRouting(router);
-  const { clearCurrentSession } = useSessionStorage();
-  const { setCurrentToken } = useSessionToken();
+  const { clearCurrentSession, setCurrentToken } = useSessionStorage();
 
   const selectFeatureToggles = useMemo(makeSelectFeatureToggles, []);
   const { isEmergencyContactEnabled } = useSelector(selectFeatureToggles);
